+++ go.mod
@@ -4,11 +4,11 @@ go 1.18
 
 require (
 	github.com/sirupsen/logrus v1.8.1
-	github.com/smacker/go-tree-sitter v0.0.0-20220323032108-9170be4a682c
+	github.com/smacker/go-tree-sitter v0.0.0-20220421092837-ec55f7cfeaf4
 	golang.org/x/exp v0.0.0-20220312040426-20fd27f61765
 )
 
 require (
-	golang.org/x/sys v0.0.0-20220403205710-6acee93ad0eb // indirect
+	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
